+++ enginecore/enginecore/state/api/outlet.py
@@ -1,0 +1,6 @@
+from enginecore.state.api.state import IStateManager
+
+
+class IOutletStateManager(IStateManager):
+    """Exposes state logic for Outlet asset """
+
