+++ enginecore/enginecore/state/api/pdu.py
@@ -1,0 +1,6 @@
+from enginecore.state.api.state import IStateManager
+
+
+class IPDUStateManager(IStateManager):
+    """Handles state logic for PDU asset """
+
